+++ lib/Doctrine/ODM/PHPCR/UnitOfWork.php
@@ -437,13 +437,15 @@ class UnitOfWork
         }
 
         foreach ($nodes as $node) {
-            $id = $node->getPath();
-            if (!isset($documents[$id])) {
+            $id       = $node->getPath();
+            $document = $this->getDocumentById($id) ?: (isset($documents[$id]) ? $documents[$id] : null);
+
+            if (! $document) {
                 continue;
             }
 
-            $document = $documents[$id];
-            $class = $this->dm->getClassMetadata(get_class($document));
+            $documents[$id] = $document;
+            $class          = $this->dm->getClassMetadata(get_class($document));
 
             $documentState = array();
             $nonMappedData = array();
@@ -720,12 +722,22 @@ class UnitOfWork
         $this->doBindTranslation($document, $locale, $class);
     }
 
-    private function doBindTranslation($document, $locale, ClassMetadata $class, $supressEvents = false)
+    /**
+     * @param object        $document
+     * @param string        $locale
+     * @param ClassMetadata $class
+     */
+    private function doBindTranslation($document, $locale, ClassMetadata $class)
     {
-        if (!$supressEvents && $invoke = $this->eventListenersInvoker->getSubscribedSystems($class, Event::preBindTranslation)) {
+        $oid = spl_object_hash($document);
+
+        // only trigger the events if we bind a new translation
+        if (empty($this->documentTranslations[$oid][$locale])
+            && $invoke = $this->eventListenersInvoker->getSubscribedSystems($class, Event::preCreateTranslation)
+        ) {
             $this->eventListenersInvoker->invoke(
                 $class,
-                Event::preBindTranslation,
+                Event::preCreateTranslation,
                 $document,
                 new LifecycleEventArgs($document, $this->dm),
                 $invoke
@@ -734,20 +746,8 @@ class UnitOfWork
 
         $this->setLocale($document, $class, $locale);
 
-        $oid = spl_object_hash($document);
-
         foreach ($class->translatableFields as $field) {
             $this->documentTranslations[$oid][$locale][$field] = $class->reflFields[$field]->getValue($document);
-        }
-
-        if (!$supressEvents && $invoke = $this->eventListenersInvoker->getSubscribedSystems($class, Event::postBindTranslation)) {
-            $this->eventListenersInvoker->invoke(
-                $class,
-                Event::postBindTranslation,
-                $document,
-                new LifecycleEventArgs($document, $this->dm),
-                $invoke
-            );
         }
     }
 
@@ -1547,13 +1547,9 @@ class UnitOfWork
 
         if ($this->isDocumentTranslatable($class)) {
             $locale = $this->getCurrentLocale($document, $class);
-            $oid = spl_object_hash($document);
 
-            // check if the document has not been bound or if the locale has not yet been bound
-            // (note array_key_exists is necessary to handle removed translations)
-            if (empty($this->documentTranslations[$oid])
-                || !array_key_exists($locale, $this->documentTranslations[$oid])
-            ) {
+            // ensure we do not bind a previously removed translation
+            if (!$this->isTranslationRemoved($document, $locale)) {
                 $this->doBindTranslation($document, $locale, $class);
             }
         }
@@ -2229,7 +2225,6 @@ class UnitOfWork
         $this->invokeGlobalEvent(Event::postFlush, new ManagerEventArgs($this->dm));
 
         if (null === $document) {
-            $this->documentTranslations = array();
             foreach ($this->documentLocales as $oid => $locales) {
                 $this->documentLocales[$oid]['original'] = $locales['current'];
             }
@@ -2237,7 +2232,6 @@ class UnitOfWork
             $documents = is_array($document) ? $document : array($document);
             foreach ($documents as $doc) {
                 $oid = spl_object_hash($doc);
-                unset($this->documentTranslations[$oid]);
                 if (isset($this->documentLocales[$oid])) {
                     $this->documentLocales[$oid]['original'] = $this->documentLocales[$oid]['current'];
                 }
@@ -2621,7 +2615,7 @@ class UnitOfWork
                                     }
 
                                     if ($referencingNode->hasProperty($referencingField['property'])) {
-                                        if (!array_key_exists($uuid, $referencingNode->getPropertyValue($referencingField['property']))) {
+                                        if (!in_array($uuid, $referencingNode->getProperty($referencingField['property'])->getString())) {
                                             if (!$collection instanceof PersistentCollection || !$collection->isDirty()) {
                                                 // update the reference collection: add us to it
                                                 $collection->add($document);
@@ -3372,8 +3366,15 @@ class UnitOfWork
 
         $currentLocale = $this->getCurrentLocale($document, $metadata);
 
-        // if no locale is specified, we reset the current translation
-        $locale = $locale ?: $currentLocale;
+        $oid = spl_object_hash($document);
+        if (null === $locale || $locale === $currentLocale) {
+            // current locale is already loaded and not removed
+            if (!$refresh && isset($this->documentTranslations[$oid][$currentLocale])) {
+                return;
+            }
+
+            $locale = $currentLocale;
+        }
 
         if (!$refresh && $this->doLoadPendingTranslation($document, $metadata, $locale)) {
             $localeUsed = $locale;
@@ -3381,9 +3382,8 @@ class UnitOfWork
             $localeUsed = $this->doLoadDatabaseTranslation($document, $metadata, $locale, $fallback, $refresh);
         }
 
-        $this->doBindTranslation($document, $localeUsed, $metadata, true);
+        $this->doBindTranslation($document, $localeUsed, $metadata);
 
-        $oid = spl_object_hash($document);
         foreach ($metadata->translatableFields as $fieldName) {
             $this->originalData[$oid][$fieldName] = $this->originalTranslatedData[$oid][$localeUsed][$fieldName]
                 = $metadata->getFieldValue($document, $fieldName);
@@ -3456,7 +3456,12 @@ class UnitOfWork
         }
     }
 
-    private function cascadeDoLoadTranslation($document, $mapping, $locale)
+    /**
+     * @param object $document
+     * @param array  $mapping
+     * @param string $locale
+     */
+    private function cascadeDoLoadTranslation($document, array $mapping, $locale)
     {
         if (!$document || !($mapping['cascade'] & ClassMetadata::CASCADE_TRANSLATION)) {
             return;
@@ -3471,10 +3476,18 @@ class UnitOfWork
             try {
                 $this->doLoadTranslation($document, $class, $locale, true);
             } catch (\Exception $e) {
+                // do nothing
             }
         }
     }
 
+    /**
+     * @param object $document
+     * @param string $locale
+     *
+     * @throws MissingTranslationException
+     * @throws PHPCRException
+     */
     public function removeTranslation($document, $locale)
     {
         $class = $this->dm->getClassMetadata(get_class($document));
@@ -3506,6 +3519,32 @@ class UnitOfWork
         $this->setLocale($document, $metadata, null);
     }
 
+    /**
+     * Checks if the translation was removed
+     * Note it also returns true if the document isn't translated
+     * or was not translated into the given locale, ie.
+     * it does not check if there is a translation for the given locale
+     *
+     * @param object $document
+     * @param string $locale
+     *
+     * @return bool
+     */
+    private function isTranslationRemoved($document, $locale)
+    {
+        $oid = spl_object_hash($document);
+
+        return isset($this->documentTranslations[$oid])
+            && empty($this->documentTranslations[$oid][$locale])
+            && array_key_exists($locale, $this->documentTranslations[$oid])
+        ;
+    }
+
+    /**
+     * @param object        $document
+     * @param ClassMetadata $metadata
+     * @throws PHPCRException
+     */
     private function doRemoveAllTranslations($document, ClassMetadata $metadata)
     {
         if (!$this->isDocumentTranslatable($metadata)) {
@@ -3517,6 +3556,11 @@ class UnitOfWork
         $strategy->removeAllTranslations($document, $node, $metadata);
     }
 
+    /**
+     * @param object        $document
+     * @param ClassMetadata $metadata
+     * @param string        $locale
+     */
     private function setLocale($document, ClassMetadata $metadata, $locale)
     {
         if (!$this->isDocumentTranslatable($metadata)) {